move note's data to parent
