+++ client/src/components/PlayerNotes.js
@@ -12,28 +12,29 @@ const user_id = userData?.user_id;
 const BASE_URL = process.env.REACT_APP_BASE_URL || '';
 
 function PlayerNotes(props) {
-    const {playerInfo} = props;
-    const [notes, setNotes] = useState([]);
+    // const {playerInfo, notes, getPlayerNotes} = props;
+    const {playerInfo, notes} = props;
+
+    // const [notes, setNotes] = useState([]);
     const [newNote, setNewNote] = useState("")
     const {playerId} = useParams();
-
-    const getPlayerNotes = async (playerId) => {
-        try {
-            const fetchedData = await axios.get(BASE_URL + `/api/myplayers/${playerId}/${user_id}`, { headers: authHeader() });
-            console.log(fetchedData.data)
-            setNotes(fetchedData.data);
-        } catch (error) {
-            console.error(error)
-        }
-    }
-    
-    useEffect(() => {
-        getPlayerNotes(playerInfo.reference );
-    }, []);
+console.log(playerId)
+    // const getPlayerNotes = async (playerId) => {
+    //     try {
+    //         const fetchedData = await axios.get(BASE_URL + `/api/myplayers/${playerId}/${user_id}`, { headers: authHeader() });
+    //         console.log(fetchedData.data)
+    //         setNotes(fetchedData.data);
+    //     } catch (error) {
+    //         console.error(error)
+    //     }
+    // }
+    // 
+    // useEffect(() => {
+    //     getPlayerNotes(playerInfo.reference );
+    // }, []);
 
     const handleChange = (event) => {
         setNewNote(event.target.value); // Update the state when the input value changes
-
       };
 
     const handleSubmit = async () => {
@@ -41,7 +42,7 @@ function PlayerNotes(props) {
             await axios.post(BASE_URL + `/api/myplayers/${playerInfo.reference}/${user_id}/add`,
                 { newNote: newNote }, { headers: authHeader() }
             )
-                .then(getPlayerNotes(playerInfo.reference ))
+                // .then(getPlayerNotes(playerInfo.reference ))
         } catch (error) {
             console.error(error);
         }
@@ -50,7 +51,7 @@ function PlayerNotes(props) {
     const clearInput = () => {
         setNewNote("");
     }
-
+    console.log(notes)
     return (
         <>
         <Container className='d-flex justify-content-center align-items-center'>
@@ -58,11 +59,11 @@ function PlayerNotes(props) {
               Player notes
               <Card.Body>
                 <Card.Title>
-                    {notes?.notes[0].notes}
+                    {notes?.notes[0]?.notes}
                 </Card.Title>
                 {notes?.extraNotes?.map(note => (
                     <div>
-                        {note.note_content}
+                        {note?.note_content}
                     </div>
                 ))}
                 <input type="text" value={newNote} onChange={handleChange}> 
@@ -70,7 +71,7 @@ function PlayerNotes(props) {
                 <button onClick={() => {handleSubmit(); clearInput()}}>Submit</button>
               </Card.Body>
             </Card>
-            {console.log(newNote)}        
+            {console.log(notes)}        
         </Container>
         </>
     )
